changement du type pour element de competence en byte

+++ src/PlanC.EntityDataModel/CoursCompetenceElements.cs
@@ -14,7 +14,7 @@ namespace PlanC.EntityDataModel
         [Key]
         [DataType("TINYINT")]
         [Column("ElementCompetenceSQNBR")]
-        public int ElementCompetenceSqnbr { get; set; }
+        public byte ElementCompetenceSqnbr { get; set; }
         [Column("CritereElementCompetenceSQNBR")]
         [DataType("TINYINT")]
         public int? CritereElementCompetenceSqnbr { get; set; }
